Suggested Changes (#547)
