+++ psst-gui/src/data/mod.rs
@@ -61,7 +61,7 @@ pub use crate::data::{
     user::{PublicUser, UserProfile},
     utils::{Cached, Float64, Image, Page},
 };
-use crate::webapi::{TrackCredits, WebApi};
+use crate::ui::credits::TrackCredits;
 
 pub const ALERT_DURATION: Duration = Duration::from_secs(5);
 
@@ -88,7 +88,6 @@ pub struct AppState {
     pub added_queue: Vector<QueueEntry>,
     pub lyrics: Promise<Vector<TrackLines>>,
     pub credits: Option<TrackCredits>,
-    pub webapi: Arc<WebApi>,
 }
 
 impl AppState {
@@ -112,12 +111,6 @@ impl AppState {
             queue: Vector::new(),
             volume: config.volume,
         };
-        let webapi = Arc::new(WebApi::new(
-            SessionService::empty(),
-            None, // Remove config.proxy_url.as_deref()
-            None, // Remove config.cache_dir.clone()
-            config.paginated_limit,
-        ));
         Self {
             session: SessionService::empty(),
             nav: Nav::Home,
@@ -179,7 +172,6 @@ impl AppState {
             finder: Finder::new(),
             lyrics: Promise::Empty,
             credits: None,
-            webapi,
         }
     }
 }
