+++ psst-gui/src/webapi/client.rs
@@ -10,7 +10,7 @@ use std::{
 use druid::{
     im::Vector,
     image::{self, ImageFormat},
-    Data, ImageBuf, Lens,
+    Data, ImageBuf,
 };
 
 use itertools::Itertools;
@@ -34,48 +34,10 @@ use crate::{
         Playlist, PublicUser, Range, Recommendations, RecommendationsRequest, SearchResults,
         SearchTopic, Show, SpotifyUrl, Track, TrackLines, UserProfile,
     },
-    error::Error,
+    error::Error, ui::credits::TrackCredits,
 };
 
 use super::{cache::WebApiCache, local::LocalTrackManager};
-
-#[derive(Debug, Clone, Data, Lens, Deserialize)]
-pub struct TrackCredits {
-    #[serde(rename = "trackUri")]
-    pub track_uri: String,
-    #[serde(rename = "trackTitle")]
-    pub track_title: String,
-    #[serde(rename = "roleCredits")]
-    pub role_credits: Arc<Vec<RoleCredit>>,
-    #[serde(rename = "extendedCredits")]
-    pub extended_credits: Arc<Vec<String>>,
-    #[serde(rename = "sourceNames")]
-    pub source_names: Arc<Vec<String>>,
-}
-
-#[derive(Debug, Clone, Data, Lens, Deserialize)]
-pub struct RoleCredit {
-    #[serde(rename = "roleTitle")]
-    pub role_title: String,
-    pub artists: Arc<Vec<CreditArtist>>,
-}
-
-#[derive(Debug, Clone, Data, Lens, Deserialize)]
-pub struct CreditArtist {
-    pub uri: Option<String>,
-    pub name: String,
-    #[serde(rename = "imageUri")]
-    pub image_uri: Option<String>,
-    #[serde(rename = "externalUrl")]
-    pub external_url: Option<String>,
-    #[serde(rename = "creatorUri")]
-    pub creator_uri: Option<String>,
-    #[serde(default)]
-    pub subroles: Arc<Vec<String>>,
-    #[serde(default)]
-    pub weight: f64,
-}
-
 pub struct WebApi {
     session: SessionService,
     agent: Agent,
@@ -321,20 +283,6 @@ impl WebApi {
         {
             log::error!("failed to read local tracks: {}", err);
         }
-    }
-
-    pub fn get_track_credits(&self, track_id: &str) -> Result<TrackCredits, Error> {
-        let request = self
-            .get(
-                format!("track-credits-view/v0/experimental/{}/credits", track_id),
-                Some("spclient.wg.spotify.com"),
-            )?
-            .set(
-                "client-token",
-                self.session.client_token().unwrap_or_default(),
-            );
-        let response = Self::with_retry(|| Ok(request.clone().call()?))?;
-        Ok(serde_json::from_str(&response.into_string()?)?)
     }
 
     fn load_and_return_home_section(&self, request: Request) -> Result<MixedView, Error> {
@@ -999,6 +947,13 @@ impl WebApi {
         Ok(result)
     }
 
+    pub fn get_track_credits(&self, track_id: &str) -> Result<TrackCredits, Error> {
+        let request = self
+            .get(format!("track-credits-view/v0/experimental/{}/credits", track_id),Some("spclient.wg.spotify.com"))?;
+        let result: TrackCredits = self.load(request)?;
+        Ok(result)
+    }
+    
     pub fn get_lyrics(&self, track_id: String) -> Result<Vector<TrackLines>, Error> {
         #[derive(Default, Debug, Clone, PartialEq, Deserialize, Data)]
         #[serde(rename_all = "camelCase")]
