+++ psst-gui/src/webapi/mod.rs
@@ -3,4 +3,3 @@ mod client;
 mod local;
 
 pub use client::WebApi;
-pub use client::{CreditArtist, RoleCredit, TrackCredits};
